platform/x86/amd/hsmp: Move platform device specific code to plat.c
commit e47c018a0ee6962fe3dd895407e2c49538cc066d upstream.

An upcoming change splits HSMP driver into ACPI and platform device
variants. Prepare for the split by moving platform device part to plat.c.

No functinality/logical changes.
Common code which can be used by ACPI and platform device
remains in hsmp.c.
ACPI code in hsmp.c will be moved to acpi.c in next patch.

Link: https://lore.kernel.org/r/20241021111428.2676884-5-suma.hegde@amd.com
